requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
sqlparse==0.4.2
psycopg2-binary
python-decouple
djangorestframework
PyJWT==2.3.0
djangorestframework-simplejwt==4.8.0